requirements.txt
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
langdetect==1.0.8
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
parsel==1.5.2
Pillow==7.1.1
Protego==0.1.16
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dotenv==0.12.0
pytz==2019.3
queuelib==1.5.0
requests==2.23.0
Scrapy==2.0.1
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
urllib3==1.25.8
w3lib==1.21.0
wrapt==1.12.1
zope.interface==5.1.0